begin rework of card game
