+++ src/components/experiences/CardGameExperience/CardDropZone.tsx
@@ -13,7 +13,7 @@ export default function CardDropZone({
     <mesh position={cardDropZonePosition}>
       <planeGeometry args={[size.cardWidth, size.cardHeight, 1]} />
       <meshBasicMaterial color='#000' opacity={0.3} transparent={true} />
-      <Html transform sprite prepend center zIndexRange={[-1, 0]}>
+      <Html center zIndexRange={[-1, 0]}>
         <p className='max-w-[4ch] select-none font-pressStart text-[8px] text-white md:text-[8px]'>Card Drop Zone</p>
       </Html>
     </mesh>
