+++ src/components/layout/Section.tsx
@@ -1,6 +1,6 @@
 export default function Section({ children, id }: { children: React.ReactNode; id: string }) {
   return (
-    <section id={id} className='relative flex h-svh w-full flex-col'>
+    <section id={id} className='relative flex h-svh max-h-[calc(100svh-96px)] w-full flex-col'>
       {children}
     </section>
   )
